split main entrypoint into modules
